requirements.txt
torch
transformers
diffusers
accelerate
opencv-python-headless
gradio
python-dotenv